module-info for core+pg-testing. Move TenantPgPoolContainer
